Fixed conversion from size_t to int)

+++ exercises/tom_clayton/ex02_oo_basics/src/ShapeSorter.h
@@ -7,7 +7,7 @@ class ShapeSorter
 {
   private: 
     std::vector<std::shared_ptr<Shape>> m_shpVect;
-    int m_vecLen;
+    size_t m_vecLen;
     
   public:
     ShapeSorter(std::vector<std::shared_ptr<Shape>> shpVect);
